+++ CHANGELOG.md
@@ -9,6 +9,7 @@ v2.7.3
 * Improve the exception message when the class isn't found when looking for a field, method, or
   constructor.
 * Support logical shifts by longs, although only the lower 5 or 6 bits are effective.
+* Fixed an infinite loop when looking for superclasses of java.lang.Object.
 
 v2.7.2 (2026-02-23)
 ------